Add TemplateInvocatorMixin
* Naming comes from usage in docs: https://argoproj.github.io/argo-workflows/workflow-concepts/#template-invocators
